// src/interfaces/user.ts
// dati utente

export interface User{

    id: number;
    name: string;
    surname: string;
    birthdate: string;
    gender: string;
    city: string;
    province: string;
    address: string;
    telnumber: string;
    email: string;
    password: string;
    study: string;
    description: string;
    skills: string[];
    curriculum: string;
    image: string;

}