// src/interfaces/businessuser.ts
// dati di un attività
export interface BusinessUser{

    id: number;
    name: string;
    surname: string;
    birthdate: string;
    telnumber: string;
    email: string;
    password: string;
    description: string;
    businessname: string;
    businesstype: string;
    businesscity: string;
    businessprov: string;
    businessaddress: string;
    businesslat: number;
    businesslong: number;
    image: string;

}